Only display variants' associated terms in Product Edit page rather than including the parent product name, which is redundant and cluttered.
Another reason why we should not rely on the post title of the variant is because it's fixed with the product name and terms' names. If by any chance a term is modified, or the parent product title is modified, the variants' titles are not updated. We should also replace variant title in Checkout page by programmatically combine parent's title and associated terms (will do later).

Granted the work around is not pretty (caching terms in a static variable), but that will do for now. The code also makes sure the variation terms are always output in a consistent manner. Previously, it was inconsistent (on one row: Red, X ; on the other: X, Blue).

git-svn-id: https://wp-e-commerce.googlecode.com/svn/branches/3.8@1214 7bdd534d-2f5a-8b57-a18c-7e4837153298