Added staged cleanup to make sure WPEC profiles are in a reasonable state.  Cleanup happens once.
Note: This is essentially an upgrade routine but cannot be run in the normal WPEC upgrade path becuase
it is possible that total processing could take minutes to a small number of hours on slower sites.
Processing time is dependent on how many of the old style temporary profiles have accumulated
in the database.  Tests with 50,000 expired profiles on a fast highly tuned system show that
cleaning the profiles can take about half an hour.

The processing in this routine works by deleting (in batches) already expired profiles.  THen for
the profiles that remain they are converted to new style profiles by adding
the temporary_profile meta and an expire time in the future.  The normal clean up routine
will take care of the clean up when the profile ages a little bit.

Fixed some typos in customer.php comments.