Remove a hidden pass by reference
I've been getting a lot of "PHP Strict standards:  Only variables should be passed by reference in /Users/John/Sites/wpec-dev/wp-content/plugins/wp-e-commerce/wpsc-components/theme-engine-v1/helpers/template-tags.php on line 647" in my debug.log, and discovered that the pass-by-reference is happening because array_shift() is being passed the result of an explode(). This removes that, and yoda-izes a the if condition.