Removed Sale Tag and renamed Sale Price
Sale tag is ugly, too opinionated, and not very supportive of theme
development.  Additionally, the fact that an item is on sale is redudant
with sale tag, sale price, and crossed out reg price.

Renamed Price: to Sale: for Sale items to clearly designate an item is on
Sale and to be less opinionated for theme developers.